refactor: (#21) Favorites 관심사 분리
